REFACTOR: fix compilation

+++ P2P/plugin/communication/fermat-p2p-plugin-communication-network-node-bitdubai/src/main/java/com/bitdubai/fermat_p2p_plugin/layer/communications/network/node/developer/bitdubai/version_1/structure/database/jpa/daos/NodeCatalogDao.java
@@ -162,7 +162,7 @@ public class NodeCatalogDao extends AbstractBaseDao<NodeCatalog> {
 
             q.setParameter("id", identityPublicKey);
 
-            return q.getSingleResult().intValue();
+            return (long) q.getSingleResult().intValue();
 
         } catch (Exception e){
             throw new CantReadRecordDataBaseException(e, "Network Node", "");
@@ -187,7 +187,7 @@ public class NodeCatalogDao extends AbstractBaseDao<NodeCatalog> {
 
             q.setParameter("triedToPropagateTimes", currentNodesInCatalog);
 
-            return q.getSingleResult().intValue();
+            return (long) q.getSingleResult().intValue();
 
         } catch (Exception e){
             throw new CantReadRecordDataBaseException(e, "Network Node", "");
